Explorer home: fix folder-stack thumbnail contrast, simplify recents to a list
The folder bookmark card's chip stack sat on a wash nearly identical to the
chip fill, making the fanned entries hard to read. The stack well now uses the
darker base --bg and the chips are lifted toward --fg (fill and border), so
they separate cleanly in both themes.

Recent files drop the launcher-card grid for a plain list capped at 5 rows:
name, file path, and a "3d ago" last-opened stamp. timeAgo moves from the
builder's AppPreviewCard into the shared platform format helpers; the now
unused .fh-grid/.fh-card styles are removed.
