+++ frontend/src/apps/explorer/FilesHome.tsx
@@ -4,10 +4,8 @@
 // bookmarks and recent files. Entering any target navigates into
 // /explorer/view/... (the explorer proper).
 import { useLayoutEffect, useRef, useState } from "react";
-import type { ReactNode } from "react";
 import { navigate, navigateUrl } from "@platform/lib/router";
-import { basename } from "@platform/lib/format";
-import { iconForEntry } from "@platform/ui/FileIcons";
+import { basename, timeAgo } from "@platform/lib/format";
 import type { Config } from "@platform/lib/api";
 import { allBookmarks, loadBookmarks } from "@platform/lib/bookmarks";
 import { useBookmarksVersion } from "@platform/lib/hooks";
@@ -15,9 +13,9 @@ import { loadRecents, recentFsPath, useRecentsVersion } from "@apps/explorer/lib
 import { BookmarkPreviewCard } from "@apps/explorer/BookmarkCards";
 import { HeroBrand } from "@platform/ui/HeroBrand";
 
-// How many recent files earn a card. The sidebar shows a tight top-3; the
-// homepage has room to be a real jump-off point.
-const MAX_RECENT_CARDS = 8;
+// How many recent files the list shows. The sidebar shows a tight top-3; the
+// homepage list stays short too — a jump-off point, not a history browser.
+const MAX_RECENTS = 5;
 
 // How many bookmark-grid rows show before the "Show more" fold.
 const BOOKMARK_ROWS = 2;
@@ -42,25 +40,25 @@ function useGridColumns(ref: React.RefObject<HTMLDivElement | null>, mounted: bo
   return cols;
 }
 
-// A launcher card: icon tile + name + path. Shared shape for bookmarks,
-// recents, and the workspace root so the grids read as one system. An anchor
-// so middle-click / Cmd-click open a new tab (same rationale as app cards).
-function LaunchCard({
+// One recents row: name, path, last-opened stamp. An anchor so middle-click /
+// Cmd-click open a new tab (same rationale as app cards).
+function RecentRow({
   href,
-  icon,
   name,
   path,
+  openedAt,
   onOpen,
 }: {
   href: string;
-  icon: ReactNode;
   name: string;
   path: string;
+  openedAt: string;
   onOpen: () => void;
 }) {
+  const when = timeAgo(Date.parse(openedAt) / 1000);
   return (
     <a
-      className="fh-card"
+      className="fh-recent"
       href={href}
       title={path}
       onClick={(e) => {
@@ -70,13 +68,9 @@ function LaunchCard({
         onOpen();
       }}
     >
-      <span className="fh-card-icon" aria-hidden="true">
-        {icon}
-      </span>
-      <span className="fh-card-text">
-        <span className="fh-card-name">{name}</span>
-        <span className="fh-card-path">{path}</span>
-      </span>
+      <span className="fh-recent-name">{name}</span>
+      <span className="fh-recent-path">{path}</span>
+      {when && <span className="fh-recent-when">{when}</span>}
     </a>
   );
 }
@@ -96,7 +90,7 @@ export default function FilesHome({ config }: { config: Config }) {
   const shownBookmarks = expanded ? bookmarks : bookmarks.slice(0, fold);
   // Raw MRU, not the sidebar's stable-slot top-3 — a full page doesn't jump
   // under the pointer the way a always-visible sidebar section does.
-  const recents = loadRecents().entries.slice(0, MAX_RECENT_CARDS);
+  const recents = loadRecents().entries.slice(0, MAX_RECENTS);
 
   return (
     <div className="files-home">
@@ -165,17 +159,17 @@ export default function FilesHome({ config }: { config: Config }) {
         <section className="fh-section">
           <h2 className="fh-heading">Recent files</h2>
           {recents.length ? (
-            <div className="fh-grid">
+            <div className="fh-recents">
               {recents.map((r) => {
                 const fsPath = recentFsPath(r.url);
                 const name = r.title || basename(fsPath);
                 return (
-                  <LaunchCard
+                  <RecentRow
                     key={fsPath}
                     href={r.url}
-                    icon={iconForEntry(basename(fsPath), false)}
                     name={name}
                     path={fsPath}
+                    openedAt={r.openedAt}
                     onOpen={() => navigateUrl(r.url)}
                   />
                 );
