+++ frontend/src/platform/lib/format.ts
@@ -18,6 +18,22 @@ export function formatMtime(epochSeconds: number | null | undefined): string {
   return new Date(epochSeconds * 1000).toLocaleString();
 }
 
+// "3d ago" style stamp; null when no time is known.
+export function timeAgo(epochSeconds: number | null | undefined): string | null {
+  if (!epochSeconds) return null;
+  const s = Math.max(0, Date.now() / 1000 - epochSeconds);
+  if (s < 60) return "just now";
+  const m = Math.floor(s / 60);
+  if (m < 60) return `${m}m ago`;
+  const h = Math.floor(m / 60);
+  if (h < 24) return `${h}h ago`;
+  const d = Math.floor(h / 24);
+  if (d < 30) return `${d}d ago`;
+  const mo = Math.floor(d / 30);
+  if (mo < 12) return `${mo}mo ago`;
+  return `${Math.floor(mo / 12)}y ago`;
+}
+
 export function basename(fsPath: string): string {
   const parts = fsPath.split("/").filter((s) => s.length > 0);
   return parts.length ? parts[parts.length - 1] : "/";
