+++ frontend/src/styles/preferences.css
@@ -127,32 +127,57 @@
   color: var(--fg-muted);
 }
 
-.fh-grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(240px, 1fr));
-  gap: 12px;
-}
-
-/* One launcher card: icon tile on the left, name over path on the right. */
-.fh-card {
+/* Recents: a plain list — name, path, last-opened stamp per row. */
+.fh-recents {
   display: flex;
-  align-items: center;
+  flex-direction: column;
+  gap: 2px;
+}
+
+.fh-recent {
+  display: flex;
+  align-items: baseline;
   gap: 12px;
-  padding: 12px 14px;
-  background: var(--bg-alt);
-  border: 1px solid var(--border);
-  border-radius: 12px;
-  text-decoration: none;
-  color: var(--fg);
+  padding: 7px 10px;
   min-width: 0;
-  transition: border-color 120ms ease, background 120ms ease;
+  color: var(--fg);
+  text-decoration: none;
+  border-radius: 8px;
 }
 
-.fh-card:hover {
-  border-color: color-mix(in srgb, var(--accent) 45%, var(--border));
-  background: color-mix(in srgb, var(--accent) 4%, var(--bg-alt));
+.fh-recent:hover {
+  background: rgba(var(--tint), 0.05);
 }
 
+.fh-recent-name {
+  flex: 0 1 auto;
+  font-size: 13.5px;
+  font-weight: 600;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+
+.fh-recent-path {
+  flex: 1 1 auto;
+  min-width: 0;
+  font-size: 11.5px;
+  color: var(--fg-muted);
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+
+.fh-recent-when {
+  flex: 0 0 auto;
+  margin-left: auto;
+  font-family: ui-monospace, Menlo, Consolas, monospace;
+  font-size: 11px;
+  color: var(--fg-muted);
+  white-space: nowrap;
+}
+
+/* Icon tile + name-over-path text block, shared by the bookmark card heads. */
 .fh-card-icon {
   flex: 0 0 auto;
   display: flex;
@@ -283,7 +308,9 @@
   align-items: center;
   aspect-ratio: 16 / 10;
   padding: 12px 12px 0;
-  background: rgba(var(--tint), 0.04);
+  /* The base --bg (a step darker than the card's --bg-alt) so the chips —
+     lifted toward --fg below — read clearly against the well they sit in. */
+  background: var(--bg);
   border-top: 1px solid var(--border);
   overflow: hidden;
 }
@@ -296,8 +323,10 @@
   flex: 0 0 auto;
   padding: 6px 10px;
   font-size: 12px;
-  background: var(--bg-panel);
-  border: 1px solid var(--border);
+  /* Lifted a step above --bg-panel (and greyed off white in light mode) so
+     the chips separate from the stack's --bg well in both themes. */
+  background: color-mix(in srgb, var(--fg) 6%, var(--bg-panel));
+  border: 1px solid color-mix(in srgb, var(--fg) 14%, var(--border));
   border-radius: 9px;
   box-shadow: 0 1px 3px var(--shadow-sm);
   transition: margin 0.18s ease;
